Switch to mjs for webpack config; use dynamic entry point configuration instead of hard-coding.
