+++ frontend/package.json
@@ -37,6 +37,7 @@
     "eslint": "^8.30.0",
     "eslint-plugin-vue": "^9.8.0",
     "eslint-webpack-plugin": "^4",
+    "glob": "^10.2.7",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
     "gulp-clean-css": "^4",
